requirements.txt
asgiref==3.8.1
click==8.1.8
dj-database-url==2.3.0
Django==5.1.7
django-environ==0.9.0
djangorestframework==3.16.0
gunicorn==20.1.0
h11==0.14.0
psycopg2-binary==2.9.10
setuptools==78.1.0
sqlparse==0.5.3
typing_extensions==4.13.0
uvicorn==0.34.0
whitenoise==6.3.0
